BLD : make sure to not use setuptools with conda
If setuptools is used, then an egg is produced.  However,
for reasons unknown the egg + enaml + conda do not play nice
with each other.  If the egg is used, then trying to use the
enaml import-hook context manager results in searching for the
enaml files in a path relative to where ever the code is being run.
This is probably a matter of not having conda set up to re-write all
the necessary hard-coded paths in the egg.

For now, just patch the setup.py file to remove the setuptools imports.
Doing this as a patch rather than by changing setup.py so that we can
still use develop with this package.