MNT : add explicit db alias
So that we can import multiple things that use mongoengine and
independently control which host/database/post each set of
odm templates looks at.

If the alias is not explicitly specified it falls back to using the
default one which is the source of the conflicts as the connection
related state is held as module-level global state in mongoengine
it will be shared by all odm_templates in the same python process.

This change should be completely transparent to all existing code.