MNT : change default database name
Change default database name to 'default-filestore'.

This name should never be used in production (so we will never step
on the toes of a production server accidentally in testing/docs).

Both testing and docs will need to make sure that they connect to
and clean up fresh DBs if required.