ENH: Torpedoing our coverage stats...
Drumroll please...

First tests on replay!

This is a simple one... it starts replay in "normal" mode, waits
one second and asks the app to stop. It then opens replay in
"live" mode, politely waits another second and then asks it to
stop too.

I wonder how this is going to play on Travis...