PRF : do not automatically dereference fields
Tell the query sets not to automatically dereference fields, which
allows the caching logic added in the previous commit to work.

This required adding a callable class `_AsDocument` to maintain a
function-level cache of seen referenced documents.  This may eventually
move to using a global LRU cache or some other method.

This has the side effect that all events with the same event descriptor
now share the same object to represent that event descriptor at the python
layer.