PRF/API: speed up retrieve
This speeds up datum retival by a factor of 30 by greedily caching
the datum documents the first time we see a resource.  By doing this
we greatly reduce the number of sockets that need to be created.

This changes the internal api of Retrive to no longer take mongoengine
documents.  This should not matter as the public API is un-changed.