API: make Document readonly
With the caching scheme there are many shared references back to the
same underlying (mutable) object.  By making them read-only we will
prevent a whole class of impossible to track down bugs due to someone
doing a `pop` in user code and then the cache layer fallying over.

This seems like a better solution than making a copy on access.