FIX/ENH: improve timezone handling
This corrects how the event timestamps are localize to the current
timezone.  For unclear reasons, `pd.to_datetime` on a `pd.Series` loses
the timezone and returns naive time objects.  We were then localizing
this time to the current time zone which created (most of the time) a 4
or 5 hour shift.  from 2100 to 2200 the day before spring daylight
savings this results in non-existent time in EST/EDT which caused
exceptions to be raised.

This will close #59 as the default behavior is now times in the local
time zone which are tz-naive.