REF: Do not user UserParameter. Unpin intake.
The release of intake 0.4.3 broke our usage of `UserParameter`, which
was never quite correct but happened to work. Our choices are to update
our usage or to drop it altogether and just use `args`. The latter is
simpler, and `UserParameter` was not adding anything in our case.

This PR thus fixes intake-bluesky for intake>=0.4.3 and unpins the
dependency.