Refactor tests.
Instead of creating a separate fixture for each catalog, put multiple
sub-catalogs in the "sources" block in the intake config file that
represents different views on the same underlying data (database or
files, as the case may be).

This seems conceptually simpler and reduces code duplication
considerably. It also steps around the immediate issue of the two
fixtures interfering with each other somehow.