ENH: make pickable without also making dottable
From the previous commit we learned that we needed to add an extra
slot to the class to let Python know that we need to go through the
custom __setstate__.  Instead of setting __dict__ to self (which makes
the dictionary "dottable" use a bool flag.  In `__setitem__` if we
don't have that attribute then we know we are in the un-pickle
path so let our selves be mutated).