FIX: if a handler is not lazy we have unknown shape don't force it
Passing the -1 in the shape through will result in an eventual shape
miss-match when trying to retrieve the data.  Rather than fail,
aggressively pull the data up.  The assumption is that if there were a
handler that was expected to load large volumes of data the handler
would be natively delayed.